tarea: se terminan componentes de tipo documento

+++ src/app/components/maestros/tipo-documento/tipo-documento.service.ts
@@ -3,46 +3,67 @@ import { AppConstants as Constant } from '../../../shared/routing/app.constants'
 import { TipoDocumento } from 'src/app/models/terceros/tipo-documento.model';
 import { AppService } from 'src/app/shared/app.service';
 import { UiService } from 'src/app/shared/ui.service';
-import { Subscription } from 'rxjs';
+import { Router } from '@angular/router';
+
 
 @Injectable()
 export class TipoDocumentoService {
 
-  listTipos: TipoDocumento[];
-  private listSub: Subscription;
+  private $tipo: TipoDocumento;
 
-  constructor(private appService: AppService, private uiService: UiService) { }
+  constructor(private appService: AppService, private uiService: UiService, private router: Router) { }
 
-  getAll(): TipoDocumento[] {
-    this.listTipos = [];
-    this.listSub = this.appService.getRequest(Constant.PATH_TIPO_DOCUMENTO).subscribe(
-      result => this.listTipos = result.body, err => console.log(err)
-    );
-    return this.listTipos;
+  fetchAll() {
+    return this.appService.getRequest(Constant.PATH_TIPO_DOCUMENTO);
   }
 
-  getById(id: number): TipoDocumento {
-    let tipo: TipoDocumento;
+  getById(id: number | string) {
     this.appService.getRequest(`${Constant.PATH_TIPO_DOCUMENTO}/${id}`).toPromise()
-      .then(res => tipo = res.body)
+      .then(res => this.$tipo = res.body)
       .catch(err => this.uiService.showSnackBar(err.message, 3));
-    return tipo;
+  }
+
+  get tipo(): TipoDocumento {
+    return this.$tipo;
+  }
+
+  returnToList() {
+    const data = {
+      title: '¿Cancelar progreso?',
+      message: 'Si vuelves perderás los avances del formulario de ingreso',
+      confirm: 'Sí, deseo regresar'
+    };
+    const dialogRef = this.uiService.showConfirm(data);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.router.navigate(['/tipo-documento']);
+      }
+    });
   }
 
   create(tipo: TipoDocumento) {
     this.uiService.putSnackBar(this.appService.postRequest(Constant.PATH_TIPO_DOCUMENTO, tipo));
+    this.router.navigate(['/tipo-documento']);
   }
 
-  update(id: number, tipo: TipoDocumento) {
+  update(id: string, tipo: TipoDocumento) {
     this.uiService.putSnackBar(this.appService.patchRequest(`${Constant.PATH_TIPO_DOCUMENTO}/${id}`, tipo));
+    this.router.navigate(['/tipo-documento']);
   }
 
-  delete(id: number) {
-    this.uiService.putSnackBar(this.appService.deleteRequest(`${Constant.PATH_TIPO_DOCUMENTO}/${id}`));
-  }
+  delete(id: string) {
+    const data = {
+      title: 'Estás seguro de eliminar el Tipo de documento?',
+      message: 'Esta acción es irreversible. \n¿Estás seguro?',
+      confirm: 'Sí, Eliminar Tipo'
+    };
+    const dialogRef = this.uiService.showConfirm(data);
 
-  cancelSubscription() {
-    if (this.listSub) { this.listSub.unsubscribe(); }
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.uiService.putSnackBar(this.appService.deleteRequest(`${Constant.PATH_TIPO_DOCUMENTO}/${id}`));
+      }
+    });
   }
 
 }
